requirements.txt
# Browser automation
playwright>=1.40.0

# LLM APIs
google-generativeai>=0.3.0
anthropic>=0.18.0
openai>=1.12.0

# Image/Video processing
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.24.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
aiofiles>=23.0.0
